+++ CHANGELOG.md
@@ -4,6 +4,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/).
 
 ### Unreleased
 
+### [8.1.0] - 2025-02-03
+
+- FEATURE: the index name can now be specified in the config
+- style(es6): replace plugin with 'this'
+
 ### [8.0.4] - 2025-02-03
 
 - dep(eslint): upgraded to v9
@@ -96,3 +101,4 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/).
 [8.0.2]: https://github.com/haraka/haraka-plugin-elasticsearch/releases/tag/v8.0.2
 [8.0.3]: https://github.com/haraka/haraka-plugin-elasticsearch/releases/tag/v8.0.3
 [8.0.4]: https://github.com/haraka/haraka-plugin-elasticsearch/releases/tag/v8.0.4
+[8.1.0]: https://github.com/haraka/haraka-plugin-elasticsearch/releases/tag/v8.1.0